+++ slang/bin/help.dart
@@ -1,0 +1,5 @@
+import 'slang.dart' as main_runner;
+
+void main(List<String> arguments) async {
+  main_runner.main(['help', ...arguments]);
+}